+++ env/index.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  exports.env = require('./dev.env.json');
+} else {
+  exports.env = require('./prod.env.json');
+}