+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## 0.4.0
-* alpha release
+* dom & user services prepared for alpha release
+
+## 0.4.1
+* changed names for dom command